+++ duckscript_sdk/src/sdk/std/test/mod.rs
@@ -1,0 +1,15 @@
+mod assert;
+
+use crate::utils::pckg;
+use duckscript::types::command::Commands;
+use duckscript::types::error::ScriptError;
+
+static PACKAGE: &str = "test";
+
+pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptError> {
+    let package = pckg::concat(parent, PACKAGE);
+
+    commands.set(assert::create(&package))?;
+
+    Ok(())
+}